# src/modules/campaign-report/dtos/responses/aggregated-events.res.dtos.ts
import { PaginatedDto } from 'src/common/dtos/offset-pagination';
import { AggregatedEventDto } from '../aggregated-event.dto';

export abstract class PaginatedEventsResponse extends PaginatedDto<AggregatedEventDto> {}