# swagger.json
{"openapi":"3.0.0","paths":{"/campaign-reports":{"get":{"operationId":"CampaignReportController_getCampaignReports","parameters":[{"name":"take","required":false,"in":"query","min":1,"max":1000,"int":true,"schema":{"minimum":1,"maximum":1000,"default":50,"type":"number"}},{"name":"page","required":false,"in":"query","min":1,"int":true,"schema":{"minimum":1,"default":1,"example":1,"type":"number"}},{"name":"order","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Order"}},{"name":"event_name","required":true,"in":"query","schema":{"example":"install | purchase","type":"string"}},{"name":"to_date","required":true,"in":"query","schema":{"example":"YYYY-MM-DD HH:00:00","type":"string"}},{"name":"from_date","required":true,"in":"query","schema":{"example":"YYYY-MM-DD HH:00:00","type":"string"}}],"responses":{"200":{"description":"Successful response with paginated campaign reports","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedEventsResponse"}}}},"400":{"description":"[4XX] Bad Request - Invalid parameters"},"500":{"description":"Internal Server Error - Unable to fetch reports"}},"summary":"Get Campaign Reports","tags":["Campaign Reports"]},"post":{"operationId":"CampaignReportController_initiateDataFetch","parameters":[],"requestBody":{"required":true,"description":"Details for fetching campaign reports within a date range","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FetchCampaignReportsReqDto"}}}},"responses":{"200":{"description":"Successful data fetch initiation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManualFetchResDto"}}}},"400":{"description":"Bad Request - Invalid payload"},"500":{"description":"Internal Server Error - Failed to fetch campaign reports"}},"summary":"Initiate Campaign Data Fetch","tags":["Campaign Reports"]}}},"info":{"title":"Impulse Test Docs","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Order":{"type":"string","enum":["asc","desc"]},"OffsetPaginationDto":{"type":"object","properties":{"take":{"type":"number"},"currentPage":{"type":"number"},"nextPage":{"type":"number"},"previousPage":{"type":"number"},"totalRecords":{"type":"number"},"totalPages":{"type":"number"}},"required":["take","currentPage","nextPage","previousPage","totalRecords","totalPages"]},"PaginatedEventsResponse":{"type":"object","properties":{"data":{"type":"array","items":{"type":"string"}},"pagination":{"$ref":"#/components/schemas/OffsetPaginationDto"}},"required":["data","pagination"]},"FetchCampaignReportsReqDto":{"type":"object","properties":{"fromDate":{"type":"string","example":"YYYY-MM-DD HH:00:00"},"toDate":{"type":"string","example":"YYYY-MM-DD HH:00:00"},"eventName":{"type":"string","example":"install | purchase"}},"required":["fromDate","toDate","eventName"]},"ManualFetchResDto":{"type":"object","properties":{"totalInserted":{"type":"number","example":25},"totalSkipped":{"type":"number","example":176},"totalFailed":{"type":"number","example":201}},"required":["totalInserted","totalSkipped","totalFailed"]}}}}